+++ src/PhpWord/Writer/HTML/Element/Image.php
@@ -18,7 +18,7 @@
 namespace PhpOffice\PhpWord\Writer\HTML\Element;
 
 use PhpOffice\PhpWord\Element\Image as ImageElement;
-use PhpOffice\PhpWord\Settings;
+use PhpOffice\PhpWord\Shared\ZipArchive;
 use PhpOffice\PhpWord\Writer\HTML\Style\Image as ImageStyleWriter;
 
 /**
@@ -77,8 +77,7 @@ class Image extends Text
             $source = substr($source, 6);
             list($zipFilename, $imageFilename) = explode('#', $source);
 
-            $zipClass = Settings::getZipClass();
-            $zip = new $zipClass();
+            $zip = new ZipArchive();
             if ($zip->open($zipFilename) !== false) {
                 if ($zip->locateName($imageFilename)) {
                     $zip->extractTo($this->parentWriter->getTempDir(), $imageFilename);