+++ CHANGELOG.md
@@ -45,6 +45,7 @@ This release marked the change of PHPWord license from LGPL 2.1 to LGPL 3; new r
 - Writer: Refactor writer parts using composite pattern - @ivanlanin
 - Docs: Show code quality and test code coverage badge on README
 - Style: Change behaviour of `set...` function of boolean properties; when none is defined, assumed true - @ivanlanin
+- Shared: Unify PHP ZipArchive and PCLZip features into PhpWord ZipArchive - @ivanlanin
 
 ## 0.10.0 - 4 May 2014